requirements.txt
Django~=4.2.1
django-ckeditor~=6.5.1